moved testCreateFreestyleProject to FreestyleProjectTest
