+++ src/test/java/school/redrover/CreateFreestyleProjectTest.java
@@ -4,33 +4,12 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
-
-import java.util.concurrent.TimeUnit;
 
 public class CreateFreestyleProjectTest extends BaseTest {
 
     private static final By DASHBOARD_BUTTON = By.linkText("Dashboard");
-
-    @Test
-    public void testCreateFreestyleProject() {
-        String expectedProjectName = "Project Test";
-
-        getDriver().findElement(By.linkText("New Item")).click();
-        getDriver().manage().timeouts().implicitlyWait(2000, TimeUnit.MILLISECONDS);
-        getDriver().findElement(By.xpath("//*[@id='name']")).sendKeys("Test");
-        getDriver().findElement(By.xpath("//*[@class='hudson_model_FreeStyleProject']")).click();
-        getDriver().findElement(By.xpath("//*[@class='btn-decorator']")).click();
-        getDriver().manage().timeouts().implicitlyWait(2000, TimeUnit.MILLISECONDS);
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-        getDriver().manage().timeouts().implicitlyWait(2000, TimeUnit.MILLISECONDS);
-
-        String actualProjectName = getDriver().findElement(By.xpath("//h1")).getText();
-
-        Assert.assertEquals(actualProjectName, expectedProjectName);
-    }
 
     @Test
     public void testCreateFreestyleProjectWithInvalidName() {
